+++ packages/react-vapor/src/components/editor/CodeEditor.tsx
@@ -42,13 +42,13 @@ const mapStateToProps = (state: IReactVaporState, {collapsibleId}: ICodeEditorPr
     isCollapsibleExpanded: CollapsibleSelectors.isExpanded(state, collapsibleId),
 });
 
-const mapDispatchToProps = (dispatch: IDispatch, {id}: {id?: string}) => ({
+const mapDispatchToProps = (dispatch: IDispatch, {id}: ICodeEditorProps) => ({
     updateStoreValue: (value: string) => dispatch(CodeEditorActions.updateValue(id, value)),
     clearCodeEditorFromStore: () => dispatch(CodeEditorActions.remove(id)),
 });
 
 class CodeEditorDisconnect extends React.Component<
-    ICodeEditorProps & Partial<ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps>>,
+    ICodeEditorProps & ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps>,
     CodeEditorState
 > {
     static defaultProps: Partial<ICodeEditorProps> = {
@@ -70,14 +70,10 @@ class CodeEditorDisconnect extends React.Component<
     private codemirror = React.createRef<ReactCodeMirror.Controlled>();
     private editor: CodeMirror.Editor;
 
-    constructor(props: ICodeEditorProps & Partial<ReturnType<typeof mapStateToProps>>, state: CodeEditorState) {
-        super(props, state);
-
-        this.state = {
-            value: props.value,
-            numberOfRefresh: 0,
-        };
-    }
+    state = {
+        value: this.props.value,
+        numberOfRefresh: 0,
+    };
 
     componentDidMount() {
         this.props.onMount?.(this.codemirror.current);